添加天狼星活动相关接口

+++ corgi-user-api/src/main/java/com/corgi/user/entity/TlxActivity.java
@@ -12,7 +12,7 @@ public class TlxActivity implements Serializable {
     private String headerImage;
     private String posterImage;
     private String meetingPoint;
-    private Integer days;
+    private Double days;
     private String departdate;
     private String period;
     private String status;
